Don't predict body init
Client doesn't handle predicted entity spawning so the organs hang around.
