+++ Content.Client/Body/Systems/BodySystem.cs
@@ -1,7 +1,13 @@
-﻿using Content.Shared.Body.Systems;
+﻿using Content.Shared.Body.Components;
+using Content.Shared.Body.Prototypes;
+using Content.Shared.Body.Systems;
 
 namespace Content.Client.Body.Systems;
 
 public sealed class BodySystem : SharedBodySystem
 {
+    protected override void InitBody(BodyComponent body, BodyPrototype prototype)
+    {
+        return;
+    }
 }
